Fixes for Visual C++ 4.0 on Windows 95 OSR
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3313 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_net.c
@@ -6267,7 +6267,7 @@ mx_network_send_client_version_time( MX_RECORD *server_record )
 
 	mx_status = mx_network_set_option( server_record,
 			MX_NETWORK_OPTION_CLIENT_VERSION_TIME | MXE_QUIET,
-			client_mx_version_time );
+			(unsigned long) client_mx_version_time );
 
 	/* Do _NOT_ return an error if the MX server sent us
 	 * an MXE_ILLEGAL_ARGUMENT error code.  This just
